minor refactorings of edge-to-edge fix
Some refactorings to @RodAlc24's (Rodrigo Alcántara) proposed fix that handles edge-to-edge enforcement in Sdk 35+:

- relocating repeated code as an extension function to AppCompatActivity into separate file
- fixing Lint error related to view inflation in PasswordStore.kt
