+++ app/src/main/java/app/passwordstore/ui/crypto/PasswordCreationActivity.kt
@@ -17,15 +17,10 @@ import android.text.InputType
 import android.view.Menu
 import android.view.MenuItem
 import android.view.View
-import android.view.ViewGroup
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.activity.result.contract.ActivityResultContracts.StartActivityForResult
 import androidx.core.content.edit
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowCompat
-import androidx.core.view.WindowInsetsCompat
 import androidx.core.view.isVisible
-import androidx.core.view.updateLayoutParams
 import androidx.core.widget.doAfterTextChanged
 import androidx.lifecycle.lifecycleScope
 import app.passwordstore.R
@@ -38,6 +33,7 @@ import app.passwordstore.util.autofill.AutofillPreferences
 import app.passwordstore.util.extensions.asLog
 import app.passwordstore.util.extensions.base64
 import app.passwordstore.util.extensions.commitChange
+import app.passwordstore.util.extensions.enableEdgeToEdgeView
 import app.passwordstore.util.extensions.getString
 import app.passwordstore.util.extensions.isInsideRepository
 import app.passwordstore.util.extensions.snackbar
@@ -147,22 +143,11 @@ class PasswordCreationActivity : BasePGPActivity() {
 
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
-    WindowCompat.enableEdgeToEdge(window)
     supportActionBar?.setDisplayHomeAsUpEnabled(true)
     title =
       if (editing) getString(R.string.edit_password) else getString(R.string.new_password_title)
     with(binding) {
-      ViewCompat.setOnApplyWindowInsetsListener(root) { v, windowInsets ->
-        val insets = windowInsets.getInsets(WindowInsetsCompat.Type.systemBars())
-        v.updateLayoutParams<ViewGroup.MarginLayoutParams> {
-          topMargin = insets.top
-          leftMargin = insets.left
-          bottomMargin = insets.bottom
-          rightMargin = insets.right
-        }
-
-        WindowInsetsCompat.CONSUMED
-      }
+      enableEdgeToEdgeView(root)
       setContentView(root)
       generatePassword.setOnClickListener { generatePassword() }
       otpImportButton.setOnClickListener {
