+++ app/src/main/java/app/passwordstore/ui/pgp/PGPKeyChangePassphraseActivity.kt
@@ -9,12 +9,7 @@ import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
 import android.view.View
-import android.view.ViewGroup
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowCompat
-import androidx.core.view.WindowInsetsCompat
-import androidx.core.view.updateLayoutParams
 import androidx.core.widget.doOnTextChanged
 import app.passwordstore.R
 import app.passwordstore.crypto.KeyUtils.tryGetId
@@ -22,6 +17,7 @@ import app.passwordstore.crypto.PGPIdentifier
 import app.passwordstore.crypto.PGPKeyManager
 import app.passwordstore.data.crypto.CryptoRepository
 import app.passwordstore.databinding.PgpKeyChangePassphraseActivityBinding
+import app.passwordstore.util.extensions.enableEdgeToEdgeView
 import app.passwordstore.util.extensions.getString
 import app.passwordstore.util.extensions.viewBinding
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
@@ -42,7 +38,6 @@ class PGPKeyChangePassphraseActivity : AppCompatActivity() {
 
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
-    WindowCompat.enableEdgeToEdge(window)
     supportActionBar?.setDisplayHomeAsUpEnabled(true)
     title = getString(R.string.pgp_change_passphrase_title)
 
@@ -54,17 +49,7 @@ class PGPKeyChangePassphraseActivity : AppCompatActivity() {
       }
 
     with(binding) {
-      ViewCompat.setOnApplyWindowInsetsListener(root) { v, windowInsets ->
-        val insets = windowInsets.getInsets(WindowInsetsCompat.Type.systemBars())
-        v.updateLayoutParams<ViewGroup.MarginLayoutParams> {
-          topMargin = insets.top
-          leftMargin = insets.left
-          bottomMargin = insets.bottom
-          rightMargin = insets.right
-        }
-
-        WindowInsetsCompat.CONSUMED
-      }
+      enableEdgeToEdgeView(root)
       setContentView(root)
       userid.text = cryptoRepository.getUserIdFromKeyId(identifier)
       oldPassphrase.doOnTextChanged { _, _, _, _ -> oldPassphraseInputLayout.error = null }
